style: move urls in one file
