+++ backend/api/views/__init__.py
@@ -1,7 +1,10 @@
-from .views import *  # noqa: W0611 (ignore unused imports warning)
+from .cseViews import *  # noqa: W0611 (ignore unused imports warning)
 from .authViews import *  # noqa: W0611
 from .courseViews import *  # noqa: W0611
 from .exerciseViews import *  # noqa: W0611
 from .resultsViews import *  # noqa: W0611
 from .sessionViews import *  # noqa: W0611
 from .studentGroupViews import *  # noqa: W0611
+from .submissionViews import *  # noqa: W0611
+from .testResultViews import *  # noqa: W0611
+from .testViews import *  # noqa: W0611
